// src/schemas/genres.ts
export enum Genre {
  ALL = 'All',
  ACTION = 'Action',
  ADVENTURE = 'Adventure',
  COMEDY = 'Comedy',
  DRAMA = 'Drama',
  HORROR = 'Horror',
  THRILLER = 'Thriller',
  SCI_FI = 'Science-Fiction',
  FANTASY = 'Fantasy',
  MYSTERY = 'Mystery',
  CRIME = 'Crime',
  ROMANCE = 'Romance',
  FAMILY = 'Family',
  ANIMATION = 'Animation',
  DOCUMENTARY = 'Documentary',
  HISTORY = 'History',
  WAR = 'War',
  WESTERN = 'Western',
  MUSIC = 'Music',
  SPORTS = 'Sports',
  BIOGRAPHY = 'Biography',
  ADULT = 'Adult',
  ANIME = 'Anime',
  CHILDREN = 'Children',
  DIY = 'DIY',
  ESPIONAGE = 'Espionage',
  FOOD = 'Food',
  LEGAL = 'Legal',
  MEDICAL = 'Medical',
  NATURE = 'Nature',
  SUPERNATURAL = 'Supernatural',
  TRAVEL = 'Travel',
  UNKNOWN = 'Unknown',
}